+++ all-is-cubes/src/inv.rs
@@ -3,6 +3,8 @@
 //! TODO: This module needs a better name; I'd be calling it `inventory` if that weren't
 //! also the name of one of its internal modules.
 
+mod icons;
+pub use icons::*;
 mod inventory;
 pub use inventory::*;
 mod tool;