+++ all-is-cubes/src/inv/icons.rs
@@ -23,8 +23,10 @@ use crate::util::YieldProgress;
 
 /// Blocks that are icons for [`Tool`]s.
 ///
-/// TODO: Consider splitting the tools part from the UI part, especially as tool icons may
-/// end up being game-specific?
+/// TODO: Should this be considered strictly part of the UI/content and not fundamentals,
+/// since it is making lots of aesthetic decisions?
+/// If so, then [`Tool::icon()`] needs to go away, and the UI will need to either contain
+/// these icons or accept them as configuration.
 #[derive(Copy, Clone, Debug, Eq, Hash, PartialEq, Exhaust)]
 #[non_exhaustive]
 pub enum Icons {